Minor refactor and cleanup
Minor refactor and cleanup

+++ BmiCalc/app/src/main/java/pamo/bmicalc/ui/calculate/CalculateViewModel.java
@@ -11,26 +11,26 @@ import java.util.Map;
 import java.util.Set;
 
 public class CalculateViewModel extends ViewModel {
-    public MutableLiveData<Map<String, Integer>> dataStore = new MutableLiveData<>();
-    public Map<String, Integer> dataMap = new HashMap<>();
+    public MutableLiveData<Map<String, Double>> dataStore = new MutableLiveData<>();
+    public Map<String, Double> dataMap = new HashMap<>();
 
     public CalculateViewModel() {
         dataStore.setValue(dataMap);
     }
 
-    public Map<String, Integer> exportAllData(){
+    public Map<String, Double> exportAllData(){
         return dataStore.getValue();
     }
 
-    public void importAllData(Map<String, Integer> newDataMap) {
+    public void importAllData(Map<String, Double> newDataMap) {
         this.dataStore.setValue(newDataMap);
     }
 
-    public Integer getMapValueByKey(String key){
+    public Double getMapValueByKey(String key){
         return dataStore.getValue().get(key);
     }
 
-    public void addDataToMap(String key, Integer value){
+    public void addDataToMap(String key, Double value){
         dataStore.getValue().put(key, value);
     }
